requirements.txt
amqp==2.6.0
asgiref==3.2.7
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.0.7
django-admin-rangefilter==0.6.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-timezone-field==4.0
feedparser==5.2.1
flower==0.9.4
future==0.18.2
httplib2==0.9.2
humanize==0.5.1
hyperlink==17.3.1
idna==2.9
importlib-metadata==1.6.1
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.10
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
numpy==1.18.5
oauthlib==2.0.6
PAM==0.4.2
pandas==1.0.4
psycopg2==2.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pyfiglet==0.8.post1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
sqlparse==0.3.1
ssh-import-id==5.7
systemd-python==234
tornado==6.0.4
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
zope.interface==4.3.2
